added a few tasks

+++ src/Algo/Tree/DFS/TreeNodesSum.cs
@@ -3,14 +3,14 @@ namespace Algo.Tree.DFS;
 //437. Path Sum III
 public class TreeNodesSum
 {
-    public int PathSum(TreeNode root, int targetSum)
+    public int PathSum(TreeNode root, long targetSum)
     {
         if (root == null) return 0;
 
         return CalcSum(root, targetSum) + PathSum(root.left, targetSum) + PathSum(root.right, targetSum);
     }
 
-    private int CalcSum(TreeNode node, int target)
+    private int CalcSum(TreeNode node, long target)
     {
         if (node == null) return 0;
